+++ packages/spl-utils/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.1](https://github.com/StrataFoundation/strata/compare/v2.1.0...v2.1.1) (2022-01-12)
+
+**Note:** Version bump only for package @strata-foundation/spl-utils
+
+
+
+
+
+# [2.1.0](https://github.com/StrataFoundation/strata/compare/v2.0.6...v2.1.0) (2022-01-11)
+
+**Note:** Version bump only for package @strata-foundation/spl-utils
+
+
+
+
+
 # [2.0.0](https://github.com/StrataFoundation/strata/compare/v1.3.2...v2.0.0) (2021-12-24)
 
 **Note:** Version bump only for package @strata-foundation/spl-utils